requirements.txt
flask>=1.1.2
gevent>=1.4.0
envparse
annoy
yara-python==4.0.1
lief>=0.9.0
tqdm>=4.31.0
numpy>=1.16.3
pandas>=0.24.2
lightgbm>=2.2.3
scikit-learn>=0.20.3
bayesian-optimization==1.2.0
PyYAML~=5.3.1
xgboost~=1.1.1
matplotlib~=3.2.2
setuptools~=47.3.1
pefile==2019.4.18
requests==2.24.0
